Title: 사각형 채우기 3, Time: 134ms, Memory: 10MB, Status: Passed - Codetree

+++ 260323/사각형 채우기 3/rectangle-fill-3.java
@@ -11,7 +11,7 @@ public class Main {
         Scanner sc = new Scanner(System.in);
         int n = sc.nextInt();
 
-        int[] memo = new int[1001];
+        long[] memo = new long[1001];
         memo[0] = 1;
         memo[1] = 2;
         for(int i = 2; i <= n; i++) {
